it just works

+++ internal/storage/postgres/model.go
@@ -1,6 +1,6 @@
 package postgres
 
 type Client struct {
-	Id     int
-	Balace int
+	Id     int64
+	Balace int64
 }
